#1002, #965: As per review, moved mapping update
functions to separate file.
